Fleshing out the Invoice class
This sets up the Invoice payment method with flags for supported
features and adds the methods that need to be implemented.

Partial capture is left out for now, although the Klarna API does
support modifying reservations. We might implement this later, but we
should focus on implementing full captures first.

Signed-off-by: Martin Svangren <martin@masv.net>